+++ lib/telegram_ex.ex
@@ -5,14 +5,16 @@ defmodule TelegramEx do
       defmodule MyBot do
         use TelegramEx, name: :my_bot
 
-        def handle_message(%{text: "/start", chat: chat}) do
-          Message.text("Hello!") |> Message.send(chat["id"])
+        def handle_message(%{text: "/start", chat: chat}, ctx) do
+          ctx
+          |> Message.text("Hello!")
+          |> Message.send(chat["id"])
         end
       end
   """
 
-  @callback handle_message(message :: map()) :: any()
-  @callback handle_callback(callback :: map()) :: any()
+  @callback handle_message(message :: map(), context :: map()) :: any()
+  @callback handle_callback(callback :: map(), context :: map()) :: any()
 
   defmacro __using__(opts) do
     quote do
@@ -21,7 +23,7 @@ defmodule TelegramEx do
       import TelegramEx
       import TelegramEx.FSM, only: [defstate: 2]
       alias TelegramEx.{API, Config, FSM}
-      alias TelegramEx.Builder.{Document, Message, Photo}
+      alias TelegramEx.Builder.{Contact, Document, Location, Message, Photo, Sticker, Video}
 
       @bot_name Keyword.fetch!(unquote(opts), :name)
       @routers Keyword.get(unquote(opts), :routers, [])
@@ -40,8 +42,8 @@ defmodule TelegramEx do
 
   defmacro __before_compile__(_env) do
     quote do
-      def handle_message(_message), do: :ok
-      def handle_callback(_callback), do: :ok
+      def handle_message(_message, _ctx), do: :ok
+      def handle_callback(_callback, _ctx), do: :ok
     end
   end
 end