+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Example project in `example/` folder as quick review all features
+
 ## [1.1.0] - 27-03-2026
 
 ### Added